+++ rarity.py
@@ -6,19 +6,20 @@ import argparse
 import key
 import commands
 import cliparser
-from transacter import Transacter
-from summoner import InvalidAmountError, InvalidSummonerError
+from transacter import Signer, Transacter
+from summoner import InvalidAddressError, InvalidAmountError, InvalidSummonerError
 
 def print_intro():
-    print(Fore.RED + 'Welcome to Rarity bot v1.3.1')
+    print('Welcome to Rarity bot v1.3.2 (devel version)')
 
-def get_address_from_args(args):
+def get_address_from_args(args, verbose = True):
     try:
         owner_address = key.load_address(args.keyfile)
     except key.InvalidInputError as e:
-        print(e)
+        print(Fore.RED + str(e) + Fore.RESET)
         exit()
-    print(Fore.WHITE + "ADDRESS FOUND, Opening " + owner_address + "\n")
+    if verbose:
+        print(Fore.WHITE + "Using address: " + owner_address + "\n")
     return owner_address
 
         
@@ -31,9 +32,15 @@ def get_private_key_from_args(args):
             # Otherwise we prompt the user to unlock the file
             private_key = key.unlock_private_key(args.keyfile)
     except key.InvalidInputError as e:
-        print(e)
+        print(Fore.RED + str(e) + Fore.RESET)
         exit()
     return private_key
+
+def get_signer_from_args(args):
+    owner_address = get_address_from_args(args, verbose = False)
+    private_key = get_private_key_from_args(args)
+    return Signer(owner_address, private_key = private_key)
+
 
 if (__name__ == "__main__"):
 
@@ -48,7 +55,7 @@ if (__name__ == "__main__"):
         try:
             key.import_new_privatekey(args.keyfile)
         except key.InvalidInputError as e:
-            print(e)
+            print(Fore.RED + str(e) + Fore.RESET)
         exit()
 
     # If keyfile doesn't exist, we exit
@@ -57,19 +64,19 @@ if (__name__ == "__main__"):
         No keyfile found. Specify it with `--keyfile path/to/keyfile.json` or save it as `privatekeyencrypted.json`.
         You can import a new key (to create a new keyfile) with the `import-key` command:
         `python3 rarity.py import-key`
-        """)
+        """ + Fore.RESET)
         exit()
            
-    # Create transacter from address and private_key (the latter only if we need to sign tx)
-    # The transacter will handle the signing and executing of transactions
-    owner_address = get_address_from_args(args)
-    if args.command in ["show", "list"]:
-        # Don't need the private key for that
-        transacter = Transacter(owner_address, private_key = None, txmode = args.txmode)
-    else:
-        private_key = get_private_key_from_args(args)
-        transacter = Transacter(owner_address, private_key = private_key, txmode = args.txmode)
+    # Create transacter (to handle calls to the blackchain) and signer (to sign tx)
+    transacter = Transacter(txmode = args.txmode)
 
+    # Check gas price
+    gas_price_gwei = transacter.get_gas_price() * 1e9
+    print("Gas price: " + str(round(gas_price_gwei, 1)) + " gwei")
+    if gas_price_gwei > args.maxgasprice:
+        print(Fore.RED + "Gas price too high (>" + \
+            str(round(args.maxgasprice, 1)) + "). Aborting." + Fore.RESET)
+        exit()
 
     # Running the main command
 
@@ -91,11 +98,19 @@ if (__name__ == "__main__"):
             commands.command_transfer(args, transacter, transfer_all = args.command == "transfer-all")
         except (InvalidAmountError, InvalidSummonerError) as e:
             print(Fore.RED + str(e))
+
+    # SEND-SUMMONER --------------
+    elif args.command == "send-summoner":
+        try:
+            commands.command_send_summoner(args, transacter)
+        except (InvalidAddressError, InvalidSummonerError) as e:
+            print(Fore.RED + e)
+
     else:
         print(Fore.RED + "Unrecognised command")    
         
     # Wait for all tx to complete (just in case!)
     transacter.wait_for_pending_transations()
-    print("\n" + Fore.RED + "Total session cost: " + str(round(transacter.session_cost, 6)) + " FTM")
+    print("\n" + Fore.RED + "Total session cost: " + str(round(transacter.session_cost, 6)) + " FTM" + Fore.RESET)